Inventory System and Item Pickups
Inventory system is a very broad use of the word here, it just holds all of the stat modifiers. But the infrastructure is here in order to make items to modify stats :).

Now we just need to add more stats and make items. Documentation will be on the doc ill write that out soon.