Refactoring old attack code
Instead of the boring rigid way done before, attacks are now done with scriptable objects! make an attack, drag it in and the character will perform the specific attack style.

Hierarchy of what can cancel what and what cant needs to be done further. Also need to clean the code more in general .-.